+++ src/console/src/store.rs
@@ -10,7 +10,7 @@ use ic_ledger_types::{BlockIndex, Tokens};
 use shared::controllers::{
     add_controllers as add_controllers_impl, remove_controllers as remove_controllers_impl,
 };
-use shared::types::interface::MissionControlId;
+use shared::types::state::MissionControlId;
 use shared::types::state::UserId;
 use shared::utils::principal_equal;
