+++ src/shared/src/types.rs
@@ -2,16 +2,16 @@ pub mod state {
     use candid::Principal;
 
     pub type UserId = Principal;
+    pub type MissionControlId = Principal;
 }
 
 pub mod interface {
     use crate::types::state::UserId;
-    use candid::{CandidType, Principal};
+    use candid::CandidType;
     use ic_ledger_types::BlockIndex;
     use serde::Deserialize;
     use std::collections::HashSet;
 
-    pub type MissionControlId = Principal;
     pub type Controllers = HashSet<UserId>;
 
     #[derive(CandidType, Deserialize)]
